requirements.txt
click==8.1.3
DateTime==4.9
Flask==2.2.2
Flask-WTF==1.0.1
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
prometheus-client==0.16.0
prometheus-flask-exporter==0.21.0
protobuf==3.20.1
psycopg2-binary==2.9.7
PyJWT==2.6.0
pymemcache==4.0.0
pytz==2022.7
uuid==1.30
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
zope.interface==5.5.2
